// frontend/typings.d.ts
export interface Student {
  stt: Number;
  _createdAt: string;
  ethnic: string;
  gender: string;
  name: string;
  address: string;
  classroom: string;
  id: string;
  day: string;
  month: string;
  year: string;
  place_of_birth: string;
  school: string;
  point_1: Number;
  point_2: Number;
  point_3: Number;
  point_4: Number;
  point_5: Number;
  total_point_1: Number;
  total_point_2: Number;
  extra_point: Number;
  phone: string;
  note: string;
}